+++ crates/release_plz_core/CHANGELOG.md
@@ -7,6 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.34.5](https://github.com/release-plz/release-plz/compare/release_plz_core-v0.34.4...release_plz_core-v0.34.5) - 2025-07-09
+
+### Other
+
+- update Cargo.toml dependencies
+
+## [0.34.4](https://github.com/release-plz/release-plz/compare/release_plz_core-v0.34.3...release_plz_core-v0.34.4) - 2025-07-01
+
+### Fixed
+
+- allow license-file in Cargo.toml ([#2287](https://github.com/release-plz/release-plz/pull/2287))
+
+## [0.34.3](https://github.com/release-plz/release-plz/compare/release_plz_core-v0.34.2...release_plz_core-v0.34.3) - 2025-06-19
+
+### Fixed
+
+- don't quit if readme doesn't exist ([#2257](https://github.com/release-plz/release-plz/pull/2257))
+
+### Other
+
+- Add some more logging related to package publishing ([#2270](https://github.com/release-plz/release-plz/pull/2270))
+- update dependencies ([#2260](https://github.com/release-plz/release-plz/pull/2260))
+- update cargo metadata to 0.20 ([#2249](https://github.com/release-plz/release-plz/pull/2249))
+
 ## [0.34.2](https://github.com/release-plz/release-plz/compare/release_plz_core-v0.34.1...release_plz_core-v0.34.2) - 2025-05-18
 
 ### Fixed